fix: split TestCoreResourceEnqueue to deal with the timeout issue
