+++ test/integration/scheduler/queueing/former/README.md
@@ -1,0 +1,2 @@
+This test file contains all the tests for the previous requeueing implementation, that is, when QueueingHint isn't implemented,
+and it is going to be cleaned up when QueueingHint feature graduates to GA.
